Use PubSubTemplate from the PubSubMessageHandler
Fixes #31

This is so the bulk of PubSub connection and message conversion logic
lives in PubSubTemplate, which can be reused by other non-SI components.

Removes unnecessary constructor.

Pass in credentials to PubSub's Publisher and Subscriber.

Right now, Google PubSub code is using the default application
credentials by default. This channels the credentials that the user
configures in GCPProperties to the GCPS API.

Code review changes

Adds "this" to member references and moves topic from constructor to
setter.