Spanner repository factory (#440)
* Make GcpConfig configuration explicitly bootstrap

It isn't AutoConfiguration so it shouldn't use those features.
There is a remaining potential issue to do with
GcpProperties (there might be 2 instances, one in bootstrap
and one in the main context, or maybe that's not an issue). Also
possibly an issue: the bootstrap configuration requires
GoogleCredentials, but isn't conditional on that class. It isn't
an issue if that class is always available, but then if it is,
why is GcpContextAutoConfiguration conditional on it?

* Avoid loading GcpContextAutoConfiguration on bootstrap

* Fix refdoc for config

* Spanner Repository and Factory Impl and Tests

* code review changes, added unit tests

* unit test change,