Create an interface for AcknowledgeablePubsubMessage (#892)
Since pulled messages contain ack IDs, but subscribed don't,
we need a common abstraction for both that simply includes
ack and nack methods without exposing the details of ack IDs.
This change accomplishes that by turning AcknowledgeablePubsubMessage
into an interface that extends a BasicAcknowledgeablePubsubMessage.
Currently, the only implementation of the interfaces is an internal PulledAcknowledgeablePubsubMessage. However, an implementation is
also envisioned on the subscribe side.