Move SQL sample tests to be under 1 database (#1135)
This moves the SQL sample tests to a single database under the spring-cloud-gcp-ci project.

The purpose of doing this is to have the properties be set in travis.yml and have the tests inherit from the application.properties that users modify. This way allows the tests to work for users who want to try out the test.

Fixes #1125.