Update binder sample properties to use topic name for input and output (#1417)
After consumer groups were integrated, the binder sample direction became misleading. It used to be that you would specify the **topic** name for output destination, and a **subscription** name for input destination. But with consumer group support, if `input.group` is missing, then an anonymous subscription will be created, with `input.destination` assumed to be the topic name. That leads to a mismatch where messages are published to TOPIC but retrieved from `anonymous.<supposed-subscription-name-which-is-actually-a-topic>.<randomUUID>`.