Replace credentials mocking with NoCredentials in GcpDatastoreAutoConfigurationTests (#2350)
Since we've run into issues with mocking from inside @Configuration classes (ByteBuddy agent hanging), might as well remove these instances of mocking credentials.

Additionally, Credentials is a class, not an interface, and my normal view is to avoid mocking classes if possible. Although Java9 Mockito uses external process attachment regardless of whether a class or an interface is mocked, so it may be less important now.