Customize project ID, user-agent string from configuration (#280)
`DatabaseClientReactiveAdapter` was already passing the configured project ID when creating `SpannerClient`. The project ID set in `SpannerOptions` is technically unused, but it seems worth explicitly setting it to the same value, so we don't have to wonder what value it is.

This PR also, since it's in the same configuration code:
* reintroduces User-Agent string for V2 (it was in a gRPC-specific implementation class in V1).
* propagates credentials. This required a small refactoring -- previously the code assumed GoogleCredentials, which precluded testing with NoCredentials.


Fixes #257.