Add PubSubTemplate ProjectTopicName/ProjectSubscriptionName parameters
The PubSubTemplate interface contract does not make it clear that it's possible to specify a fully qualified topic or subscription name, including the project. The documentation of this feature is present, but it's easily overlooked.

These convenience methods make it clear that a ProjectSubscriptionName or a ProjectTopicName can be supplied with a project, to make it easier for developers to understand this functionality/capability.