temp fix to get away from circular reference error (#713)
This is a temporarily fix to let datastore basic sample work with spring boot 2.6.
circular reference is on class org.springframework.shell.result.ResultHandlerConfig (link to error)

Suspect cause:
This sample uses spring-shell for cmd line interaction. Spring shell seems to be a project relative slow in updating. We are already on the newest version 2.0.1.RELEASE which has dependencies on spring boot 1.5.8.