fix: remove gson bean creation in spanner auto-configuration.  (#1241)
Discovered from Spring Boot Doc that auto-configuration for Gson is already provided (doc, GsonAutoConfiguration). Configuration properties are also provided for customizing.

Removing the Gson bean creation in Spanner auto-configuration in favor of this one. This bean was original created in #937 for the conversion when dealing with JSON/ARRAY Spanner type.

Note: this change is backward compatible for users.