fix(codegen): additional replacement logic in googleapis workspace file (#1489)
This PR unblocks generation script errors when pointing to more recent googleapis committish. This is due to a change in googleapis/WORKSPACE (d7e8de7) that modified this file to maven install the gapic-generator-java jar. The spring codegen workflow still relies on editing this to use a locally installed branch of gapic_generator_java. This patch adds logic to remove the maven_install() block and add back gapic_generator_java_repositories().

Note that the longer-term to address this to migrating the spring generator code out of the gapic-generator-java repo, and switching to using the gapic-generator-java jar.