chore: cleanup and remove unsed pom entries related to firestore native sample. (#1851)
This pr cleans up setups in pom for `spring-cloud-gcp-data-firestore-sample` related to native support (configured with outdated `org.springframework.experimental:spring-native-configuration`). It reverts what's added for the sample in https://github.com/GoogleCloudPlatform/spring-cloud-gcp/pull/508.
The relevant native support module has stopped working for a while and was deleted in #1287. So it should be safe to also remove these setups in sample.