fix: Don't try to access blob if autocreate is enabled
The account used to write the blob in Google Cloud Storage may not have
the necessary permissions to also read the file. In this case the
getOutputStream() implementation would be broken, and unconditionally
throw a permission error.

Fix this by not accessing the blob if the autoCreateFiles option is
enabled.