fix(doc): fix secret manager section header levels (#4060)
The table of contents for the [documentation of Spring Cloud GCP
7.1.0](https://googlecloudplatform.github.io/spring-cloud-gcp/7.1.0/reference/html/index.html)
shows '22. Global Secrets' and '23. Regional Secrets' as sibling
sections after '21. Secret Manager', but I believe they should be
subsections of 'Secret Manager', and they are also part of the
`secretmanager.adoc` file.

This pull request changes these sections to be subsections of Secret
Manager.


Fixes #4072

Co-authored-by: Nils Breunese <nbreunese@bol.com>
Co-authored-by: Min Zhu <zhumin@google.com>