chore: Provide generic Javadocs link on the Spring README (#4134)
Fixes
https://github.com/GoogleCloudPlatform/spring-cloud-gcp/issues/4113

The Javadocs page takes a bit to generate and prevents us from updating
the latest versions on the README once we kick off the Spring release.
Provide a generic link to the javadocs page that points to the latest
version by default. Users can select the version they want to use in the
drop down in the bottom right corner.