feat: add script to update versions in README.adoc (#4176)
**Description**

This PR introduces a new script to automate the process of updating the
version numbers in the README.adoc file.

**Problem**

The README.adoc file contains links to the documentation for different
versions of the project. These version numbers are hardcoded and need to
be updated manually for each new release. This is a tedious and
error-prone process.

**Solution**

This PR adds a new shell script named
`update_latest_versions_in_readme.sh` in the `docs/` folder. This
script:

1. Fetches the latest version numbers for each major release line (6.x,
5.x, 4.x, 3.x) from the maven-metadata.xml file in the Maven Central
repository.
2. Uses sed to directly update the hardcoded version numbers in the
README.adoc file.

**How to use the script**

To update the versions in the README.adoc file, simply run the following
command from the root of the project:

   1 `./docs/update_latest_versions_in_readme.sh`

This will ensure that the README.adoc file always points to the latest
released versions of the documentation.