deps(6.x): upgrade libraries-bom to 26.70.0 (#4189)
We'd love to accept your patches and contributions to this project.
There are just a few small
guidelines you need to follow before opening an issue or a PR:

1. Ensure the issue was not already reported.
2. Open a new issue if you are unable to find an existing issue
addressing your problem. Make sure
to include a title and clear description, as much relevant information
as possible, and a code
sample or an executable test case demonstrating the expected behavior
that is not occurring.
3. Discuss the priority and potential solutions with the maintainers in
the issue. The maintainers
would review the issue and add a label "Accepting Contributions" once
the issue is ready for
   accepting contributions.
4. Open a PR only if the issue is labeled with "Accepting
Contributions", ensure the PR description
clearly describes the problem and solution. Note that an open PR without
an issue labeled with "
   Accepting Contributions" will not be accepted.

See
also
[CONTRIBUTING.md](https://github.com/GoogleCloudPlatform/spring-cloud-gcp/blob/main/CONTRIBUTING.md)
.